Fix bugs in Haskell->Rust and Rust->Haskell articles
I ran into a few bugs/typos that made it impossible to directly follow
along with the articles. This fixes those for other readers in the
future.

The biggest change is compiling the C code with GHC instead of GCC,
since that seems to be the recommended practice for Haskell, and it's
what made it start working for me.